// src/types.ts
export interface ScholarshipRecord {
  lop: string; // Lớp
  khoa: string; // KHOA
  mssv: string; // Mã sinh viên
  cccd: string; // cccd
  hoLot: string; // Họ lót
  ten: string; // Tên
  ngaySinh: string; // Ngày sinh
  tenHocBong: string; // Tên học bổng
  donViTaiTro: string; // Đơn vị ban hành/tài trợ
  tienHocBong: string; // Tiền học bổng
  maHocKy: string; // Mã học kỳ
  namHoc: string; // Năm học
}